+++ test/meta_tests.jl
@@ -8,6 +8,7 @@
         "block_tests.jl" => 3,
         "c_suite_tests.jl" => 9,
         "chainrules_tests.jl" => 4,
+        "contract_tests.jl" => 3,
         "coverage_tests.jl" => 1,
         "corpus_tests.jl" => 4,
         "derivative_tests.jl" => 10,